server/llm/requirements.txt
langchain_huggingface
peft 
evaluate 
sentencepiece 
scikit-learn 
accelerate 
bitsandbytes 
transformers 
trl 
tqdm 
pandas 
matplotlib
langchain
python-dotenv